src/requirements.txt
beautifulsoup4
requests
google-cloud-firestore
spotipy